[portScope] read minval and maxval as float
svn path=/trunk/yarp2/; revision=9348

+++ src/yarpscope/src/DataPlot.h
@@ -23,8 +23,8 @@ class DataPlot : public GDatabox::Databox
 {
 public:
     DataPlot(const Glib::ustring &title,
-             int minval,
-             int maxval,
+             float minval,
+             float maxval,
              int size,
              const Glib::ustring &bgcolor,
              bool autorescale);
